backend/pyproject.toml
[project]
name = "road700-backend"
version = "0.1.0"
description = "Backend for the Road700 fleet repairs platform"
requires-python = ">=3.10"
dependencies = [
  "alembic>=1.14.0",
  "boto3>=1.35.0",
  "bcrypt<4",
  "email-validator>=2.2.0",
  "eval_type_backport>=0.3.0",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "openpyxl>=3.1.5",
  "passlib[bcrypt]>=1.7.4",
  "Pillow>=11.1.0",
  "pydantic-settings>=2.6.0",
  "psycopg[binary]>=3.2.0",
  "pypdf>=5.4.0",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.12",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
  "xlrd>=2.0.1",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
